requirements.txt
google-generativeai==0.3.2
typing==3.7.4.3